requirements.txt
pvporcupine
pvrecorder
faster-whisper
torch
TTS
sounddevice
soundfile
faiss-cpu
sentence-transformers
requests
